// aitianyu.cn/ClientApp/src/dty/model/Interfaces.ts
/**@format */

export interface IProjectStyle {
    color: string;
    fontColor: string;
}

export interface IProject {
    key: string;
    name: string;
    desc: string;
    github: string;
    repo: string;
    style: IProjectStyle;
}

export interface IBinarySource {
    name: string;
    url: string;
}

export interface IBinaries {
    [platform: string]: IBinarySource[];
}

export interface IDownload {
    key: string;
    name: string;
    github: string;
    bin: IBinaries;
}

export interface IParameters {
    [key: string]: string;
}